Restructure Submission Folder and add more assets
